requirements.txt
asgiref==3.2.7
Django==3.0.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
pandas==0.22.0
scikit-learn==0.23.2
